LPS-83331 Move things around, I really don't want this to look like a ServiceBuilder module, so I modeled it after portal-cache-api
