+++ modules/apps/document-library-opener/document-library-opener-google-drive-api/bnd.bnd
@@ -2,7 +2,5 @@ Bundle-Name: Liferay Document Library Opener Google Drive API
 Bundle-SymbolicName: com.liferay.document.library.opener.google.drive.api
 Bundle-Version: 1.0.0
 Export-Package:\
-	com.liferay.document.library.opener.google.drive.constants,\
-	com.liferay.document.library.opener.google.drive.manager,\
-	com.liferay.document.library.opener.google.drive.model,\
-	com.liferay.document.library.opener.google.drive.util
+	com.liferay.document.library.opener.google.drive,\
+	com.liferay.document.library.opener.google.drive.constants
