--- modules/apps/document-library-opener/document-library-opener-google-drive-api/src/main/resources/com/liferay/document/library/opener/google/drive/model/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
