+++ modules/apps/document-library-opener/document-library-opener-google-drive-api/src/main/resources/com/liferay/document/library/opener/google/drive/packageinfo
